Enable in-page heading/fragment links
This requires removing the fixed nav header, as it blocks the content
when the browser scrolls to the fragment.

We may be able to reintroduce the fixed nav at some point in future,
with a more sophisticated doc template, but for now, usability is
improved more by doing inline links than by fixing the top nav.

May also need to consider fixed positioning for the side nav.