Update todo-app-with-rails.md
I went through the Todo App tutorial and found a few minor inconsistencies in the docs. It refers to `app/relations/task.rb` with a class name of `Task`, which the tasks generator actually puts in `app/relations/tasks_relation.rb` with a class name of `TasksRelation`.  The file was also misnamed for the mapper.  The docs list it as being in `app/mappers/tasks.rb` when it's actual location is `app/mappers/task_mapper.rb`; its name is correct in the docs, however.  Additionally the docs mention a `done` column, which is actually named `is_completed`.  

Not entirely relevant to this changeset, but ROM looks pretty awesome!