Hide empty errors arrays in to-do app task forms
Without checking for `tasks.present?`, `full_errors` will show as an empty array (`[]`) on the task form, even when the task is fully valid.

I was confused for a moment by this, thinking it might have been some quirk of ROM's integration with Rails. Hiding the empty array makes the form appear as expected.