Invert #completed to be consistant
Later on in the tutorial !is_completed is treated as active, which
is inconsistant when is_completed is null to the relation methods.

Also added a reference to the nice docs which describe the
Relation reader methods for rom-sql. It's a nice pattern showing
how to find methods for different adapters.