Fix slashes in class names in anchored links
When the link would have an anchor (referencing a certain method) and
class name would be nested in a namespace, it would be displayed with a
slash instead of "::". E.g.

  [api::rom-repository::Repository/Root](#aggregate)

is currently displayed as

  Repository/Root#aggregate

instead of

  Repository::Root#agregate

We update the link rendering logic with the notion that the content of
links are in fact paths, not classes (maybe they were classes originally
and the logic god outdated), and we convert those paths into class names
for link content.