Add OpenGraph and Twitter Card meta tags to HTML head
These meta tags should enable social sites like Facebook and Twitter to
generate embedded card links for all the pages on the rom-rb.org site.

If a `description` field is added to a content page’s front matter, then
it will show up as the summary text in the card unit. Otherwise, the
generic ROM tagline is used.